"Event of the day" card removed. Config file updated with comments about current status.
