--- src/config/daysOfTheYearApiInfo.js
@@ -1,5 +1,0 @@
-export const daysOfTheYearApiInfo = {
-  key: "898278c921a1d2a2107a1e746f7b5592",
-  getTodaysEventUrl:
-    "https://cors-anywhere.herokuapp.com/https://www.daysoftheyear.com/api/v1/today/",
-};
